+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - [feat:support ipv6.](https://github.com/Tencent/spring-cloud-tencent/pull/1592)
 - [feat:support config all recover enabled.](https://github.com/Tencent/spring-cloud-tencent/pull/1593)
 - [feat:support stat reporting path aggregation.](https://github.com/Tencent/spring-cloud-tencent/pull/1602)
+- [feat:support namespace exports configuration if not created.](https://github.com/Tencent/spring-cloud-tencent/pull/1611)